# backend/core-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Core Service - manages cars, packages, and regions",
        "title": "Core Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8001",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/cars": {
            "get": {
                "description": "Retrieves all cars",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cars"
                ],
                "summary": "Get all cars",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core-service_pkg_response.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new car",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cars"
                ],
                "summary": "Create a new car",
                "parameters": [
                    {
                        "description": "Car data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/core-service_models_dto.CreateCarRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/core-service_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/cars/{id}": {
            "get": {
                "description": "Retrieves a specific car by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cars"
                ],
                "summary": "Get car by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Car ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core-service_pkg_response.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an existing car",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cars"
                ],
                "summary": "Update a car",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Car ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Car data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/core-service_models_dto.UpdateCarRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core-service_pkg_response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a car by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cars"
                ],
                "summary": "Delete a car",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Car ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core-service_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/packages": {
            "get": {
                "description": "Retrieves all packages",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Packages"
                ],
                "summary": "Get all packages",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core-service_pkg_response.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new package",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Packages"
                ],
                "summary": "Create a new package",
                "parameters": [
                    {
                        "description": "Package data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/core-service_models_dto.CreatePackageRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/core-service_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/packages/{id}": {
            "get": {
                "description": "Retrieves a specific package by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Packages"
                ],
                "summary": "Get package by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Package ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core-service_pkg_response.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an existing package",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Packages"
                ],
                "summary": "Update a package",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Package ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Package data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/core-service_models_dto.UpdatePackageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core-service_pkg_response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a package by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Packages"
                ],
                "summary": "Delete a package",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Package ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core-service_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/regions/provinces": {
            "get": {
                "description": "Retrieves all provinces",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Regions"
                ],
                "summary": "Get all provinces",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core-service_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/regions/provinces/{province}/regencies": {
            "get": {
                "description": "Retrieves all regencies for a specific province",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Regions"
                ],
                "summary": "Get regencies by province",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Province ID",
                        "name": "province",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core-service_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/regions/regencies/{regency}/districts": {
            "get": {
                "description": "Retrieves all districts for a specific regency",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Regions"
                ],
                "summary": "Get districts by regency",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Province ID",
                        "name": "province",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Regency ID",
                        "name": "regency",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core-service_pkg_response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "core-service_models_dto.CreateBenefitRequest": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "sortOrder": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "core-service_models_dto.CreateCarRequest": {
            "type": "object",
            "required": [
                "brand",
                "licensePlate",
                "model",
                "year"
            ],
            "properties": {
                "brand": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "licensePlate": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "transmission": {
                    "$ref": "#/definitions/models.TransmissionType"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "core-service_models_dto.CreatePackageRequest": {
            "type": "object",
            "required": [
                "name",
                "packageType",
                "price"
            ],
            "properties": {
                "benefits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core-service_models_dto.CreateBenefitRequest"
                    }
                },
                "description": {
                    "type": "string"
                },
                "discountPrice": {
                    "type": "number"
                },
                "durationMinutes": {
                    "type": "integer"
                },
                "imageUrl": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "packageType": {
                    "$ref": "#/definitions/models.PackageType"
                },
                "price": {
                    "type": "number"
                },
                "totalSessions": {
                    "type": "integer"
                }
            }
        },
        "core-service_models_dto.UpdateCarRequest": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "mileage": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.CarStatus"
                },
                "transmission": {
                    "$ref": "#/definitions/models.TransmissionType"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "core-service_models_dto.UpdatePackageRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "discountPrice": {
                    "type": "number"
                },
                "durationMinutes": {
                    "type": "integer"
                },
                "imageUrl": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "packageType": {
                    "$ref": "#/definitions/models.PackageType"
                },
                "price": {
                    "type": "number"
                },
                "status": {
                    "$ref": "#/definitions/models.PackageStatus"
                },
                "totalSessions": {
                    "type": "integer"
                }
            }
        },
        "core-service_pkg_response.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "core-service_pkg_response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/core-service_pkg_response.ErrorDetail"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "models.CarStatus": {
            "type": "string",
            "enum": [
                "available",
                "in_use",
                "maintenance"
            ],
            "x-enum-varnames": [
                "CarStatusAvailable",
                "CarStatusInUse",
                "CarStatusMaintenance"
            ]
        },
        "models.PackageStatus": {
            "type": "string",
            "enum": [
                "active",
                "inactive"
            ],
            "x-enum-varnames": [
                "PackageStatusActive",
                "PackageStatusInactive"
            ]
        },
        "models.PackageType": {
            "type": "string",
            "enum": [
                "bronze",
                "silver",
                "gold",
                "platinum"
            ],
            "x-enum-varnames": [
                "PackageTypeBronze",
                "PackageTypeSilver",
                "PackageTypeGold",
                "PackageTypePlatinum"
            ]
        },
        "models.TransmissionType": {
            "type": "string",
            "enum": [
                "manual",
                "automatic"
            ],
            "x-enum-varnames": [
                "TransmissionManual",
                "TransmissionAutomatic"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "XApiKey": {
            "type": "apiKey",
            "name": "x-api-key",
            "in": "header"
        },
        "XRequestAt": {
            "type": "apiKey",
            "name": "x-request-at",
            "in": "header"
        },
        "XServiceName": {
            "type": "apiKey",
            "name": "x-service-name",
            "in": "header"
        }
    }
}